Distributes tooltip components
